+++ basedecorater.rb
@@ -1,0 +1,11 @@
+require './nameable'
+class BaseDecorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
