+++ trimmerdecorater.rb
@@ -1,0 +1,6 @@
+require './basedecorater'
+class TrimmerDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.size > 10 ? @nameable.correct_name[0..9] : @nameable.correct_name
+  end
+end
